+++ test/formatter.test.ts
@@ -3,14 +3,7 @@ const testBinaryYaml = 'custom-binary-delims.yml';
 const testUnaryYaml = 'custom-unary-delims.yml';
 const testResultYaml = 'test-result-strings.yml';
 
-const testObject = {
-    user: 'haxzie',
-    number: 'first',
-    repo: 'wring',
-    1: 'haxzie',
-    2: 'probot',
-    3: 'wring'
-};
+const testObject = require('./test_object').testObject;
 
 test('for custom binary delimiter', () => {
     let collection = new Collection(testBinaryYaml, __dirname);
