+++ test/test_object.ts
@@ -1,0 +1,8 @@
+exports.testObject = {
+    user: 'haxzie',
+    number: 'first',
+    repo: 'wring',
+    1: 'haxzie',
+    2: 'probot',
+    3: 'wring'
+};
